requirements.txt
python-vlc>=3.0.18121
mutagen>=1.47.0
